Rename suggestions_dropdown -> combo-box
